feat(#81): Page re-rendering 최적화를 위한 컴포넌트 분리
